[refactor] csv 탭 레이아웃 수정
