+++ FE/src/components/SettingContent/CSVSetting/csvSetting.scss
@@ -1,41 +1,6 @@
+@use '../../../styles/mixins'as *;
+@use '../../../styles/values'as *;
+
 .csv__setting__container {
-  color: white;
-  display: flex;
-  justify-content: space-evenly;
-  align-items: center;
-
-  .csv__export {
-    background: transparent;
-    height: 54px;
-    color: white;
-    border: 1px 0 0 10px #219cf3 solid;
-    box-shadow: none;
-    border-radius: 16px;
-    padding: 16px;
-    overflow: visible;
-    cursor: pointer;
-    text-shadow: 0 0 10px #219cf3, 0 0 10px #219cf3;
-  }
-
-  .csv__import__form {
-    display: flex;
-    flex-direction: column;
-
-    .csv-input {
-      margin-bottom: 10px;
-    }
-
-    .csv__import {
-      background: transparent;
-      color: white;
-      width: 216px;
-      border: 1px 0 0 10px #219cf3 solid;
-      box-shadow: none;
-      border-radius: 16px;
-      padding: 16px;
-      overflow: visible;
-      cursor: pointer;
-      text-shadow: 0 0 10px #219cf3, 0 0 10px #219cf3;
-    }
-  }
+  @include settingContainer;
 }
